+++ changelog.md
@@ -1,7 +1,34 @@
+### v2.10.0
+---
+* **Version Checker**
+* The version checker can now print any detected outdated clients.
+* Simply add any argument to the chat command, e.g. `/rc v 1` to get the list.
+
+
+* **Ilvl comparisons**
+* If a trinket or ring is looted, and the candidate already has equipped a different version of the item,
+RCLootCouncil now uses that item's ilvl when calculating the difference, since that item would have to be replaced (#378).
+
+
+* The owner is now also shown above the award status in the voting frame.
+
+
+* Updated Libraries to the latest version(s).
+
+
+###### Bugfixes
+* *Sorting the loot history now properly triggers secondary sorts where needed.*
+
+###### Dev
+* Changed structure of `db.global.verTestCandidates`.
+* Replaced `:GetDiff()` with `:GetIlvlDifference()` which takes different arguments.
+
+
 ### v2.9.7
 ---
 ###### Bugfixes
 * *Fixed issue with Loot Status (#382).*
+
 
 
 ### v2.9.6